refactor(#296): extract runner control routes to routers/runners.py
Extract runner-related routes and helper functions from routers/fleet.py
into a new dedicated router module. This improves separation of concerns
and continues the god-module refactoring work from epic #159.

Changes:
- Created backend/routers/runners.py with 175 lines:
  - Runner lifecycle control endpoints: GET /api/runners, GET /api/runners/matlab,
    POST /api/runners/{runner_id}/start, POST /api/runners/{runner_id}/stop
  - Helper functions: runner_svc_path, run_runner_svc, runner_num_from_id,
    _runner_sort_key, _is_matlab_runner, _matlab_runner_summary
  - Dependency injection for required config and utilities

- Modified backend/routers/fleet.py:
  - Removed runner-related routes (now in runners.py)
  - Removed runner helper functions (now in runners.py)
  - Import runner helpers from runners module for _fleet_control_local
  - Now 137 lines (focused on fleet orchestration only)

- Modified backend/server.py:
  - Added import and router registration for runners router
  - Registered in batch-2 extracted routers section

- Fixed backend/metrics.py:
  - Updated imports to use routers.system module for shared functions
  - Fixed import of _get_runner_capacity_snapshot from routers.system

All tests pass, lint passes, no regression in runner lifecycle behavior.
